requirements.txt
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.2
Cython==3.0.11
firebase-admin==6.8.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
google-api-core==2.25.0rc0
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.0
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
libxml2-python==2.9.14
MarkupSafe==3.0.2
msgpack==1.1.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
pyzmq==26.2.0
requests==2.32.3
rsa==4.9.1
six==1.16.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.1.3
zipp==3.21.0
